pyproject.toml
[build-system]
requires = ["setuptools >= 71.0.0"]
build-backend = "setuptools.build_meta"

[project]
authors = [
    { name = "Thousand Brains Project" }
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha", # Update this status as the code evolves
    "License :: OSI Approved :: MIT License",
    "Private :: Do Not Upload to PyPI", # Remove if you want to upload to PyPI
    "Programming Language :: Python :: 3",
]
dependencies = [
    "tbp.monty", # imported via conda (thousandbrainsproject::tbp.monty)
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pillow",
    "openvr",
    "vpython",
    "opencv-python",
    # Core dependencies for visualization implementations
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
    # Plotly/Dash implementation (already included)
    "dash>=2.10.0",
    "plotly>=5.14.0",
    # Streamlit implementation
    "streamlit>=1.22.0",
    # Qt implementation
    "PyQt6>=6.4.0",
    # "PyOpenGL>=3.1.0",
    # "PyOpenGL-accelerate>=3.1.0",
    "matplotlib>=3.5.0",
    # Panel/HoloViews implementation
    "panel>=1.0.0",
    "holoviews>=1.16.0",
    "param>=1.13.0",
    "bokeh>=3.0.0",
]
description = "Monty for Ultrasound Demo - TBP Hackathon 2025"
dynamic = ["version"]
license = { file = "LICENSE" }
name = "itsnotatumah"
readme = "README.md"

[project.optional-dependencies]
dev = [
    # Check for undeclared dependencies
    "deptry",

    # Python static type checker
    "mypy==1.11.2",

    # Testing, code style, etc...
    "pytest==7.1.1",
    "pytest-xdist==2.5.0",
    "pytest-cov==3.0.0",

    # Python linter and formatter
    "ruff==0.7.1",
]

[project.urls]
Repository = "https://github.com/thousandbrainsproject/itsnotatumah"
Issues = "https://github.com/thousandbrainsproject/itsnotatumah/issues"

[tool.deptry]
exclude = ["venv", "\\.venv", "\\.direnv", "\\.git", "setup\\.py"]
experimental_namespace_package = true # recognize `tbp.` prefixed packages as a namespace package
ignore = ["DEP002"]
known_first_party = [
    "configs", # experiment configurations
]


[tool.mypy]
explicit_package_bases = true
files = ["src", "tests"]
mypy_path = ["src", "tests"]
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tbp.monty.*"
ignore_missing_imports = true # Remove once tbp.monty types are published

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -n auto"
testpaths = ["tests/unit"]
filterwarnings = [
    "ignore:.*:DeprecationWarning",
]
junit_family = "xunit1"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
preview = true
explicit-preview-rules = true # comment this out to check all preview rules
select = [
    "ALL",
    # Explicitly opt-in to preview rules
    "CPY001", # Missing copyright notice at top of file
    "DOC201", # return is not documented in docstring
    "DOC202", # Docstring should not have a returns section because the function doesn't return anything
    "DOC402", # yield is not documented in docstring
    "DOC403", # Docstring has a "Yields" section but the function doesn't yield anything
    "DOC501", # Raised exception {id} missing from docstring
    "DOC502", # Raised exception is not explicitly raised: {id}
    "E261", # Insert at least two spaces before an inline comment
    "E262", # Inline comment should start with '# '
    "E265", # Block comment should start with '# '
]
ignore = [
    "COM812", # Trailing comma missing; conflicts with ruff format
    "ISC001", # Implicitly concatenated string literals on one line; conflicts with ruff format
]

[tool.ruff.lint.per-file-ignores]
# D104: Missing docstring in public package
"__init__.py" = ["D104"]
# D100: Missing docstring in public module
# D101: Missing docstring in public class
# D102: Missing docstring in public method
# S101: Use of `assert` detected
"tests/**.py" = ["D100", "D101", "D102", "S101"]

[tool.ruff.lint.flake8-copyright]
author = "Thousand Brains Project"


[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
where = ["."]
namespaces = true
